only one do_droput implementation task-atac-denoising
